pyproject.toml
[project]
name = "reagent"
version = "0.1.0"
description = "The autonomous AI agent for binary analysis and vulnerability research."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "anthropic>=0.79.0",
    "lief>=0.17.3",
    "litellm>=1.81.12",
    "openai>=2.21.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rzpipe>=0.6.2",
    "tenacity>=9.1.4",
    "textual>=7.5.0",
    "typer>=0.23.1",
]

[project.scripts]
reagent = "reagent.cli:main"

[dependency-groups]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reagent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
